+++ packages/component/src/SendBoxToolbar/BasicSendBoxToolbar.tsx
@@ -8,7 +8,7 @@ const { useStyleOptions } = hooks;
 function BasicSendBoxToolbar({ className }: SendBoxToolbarMiddlewareProps) {
   const [{ disableFileUpload, hideUploadButton }] = useStyleOptions();
 
-  return (!disableFileUpload || !hideUploadButton) && <UploadButton className={className} />;
+  return !(disableFileUpload || hideUploadButton) && <UploadButton className={className} />;
 }
 
 export default memo(BasicSendBoxToolbar);